silentpayments: receiving
Add routine for scanning a transaction and returning the necessary
spending data for any found outputs. This function works with labels via
a lookup callback and requires access to the transaction outputs.
Requiring access to the transaction outputs is not suitable for light
clients, but light client support is enabled by exposing the
`_create_shared_secret` and `_create_output_pubkey` functions in the
API. This means the light client will need to manage their own scanning
state, so wherever possible it is preferrable to use the
`_recipient_scan_ouputs` function.

Add an opaque data type for passing around the prevout public key sum
and the input hash tweak (input_hash). This data is passed to the scanner
before the ECDH step as two separate elements so that the scanner can
multiply the scan_key * input_hash before doing ECDH.

Add functions for deserializing / serializing a prevouts_summary object to
and from a public key. When serializing a prevouts_summary object, the
input_hash is multplied into the prevout public key sum. This is so the
object can be stored as public key for wallet rescanning later, or to send
to light clients. For the light client, a `_parse` function is added which
parses the compressed public key serialization into a `prevouts_summary`
object.

Finally, add test coverage for the receiving API.